fix: keep bash detach reusable after normal completion
